+++ src/pages/onboarding/FirstMeet.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router";
 import { Button, Input, DatePickerDrawer, ProgressHeader } from "@/shared/ui";
-import { formatDate } from "@/shared/utils/date/formatdate";
+import { formatDateKorean } from "@/shared/utils";
 
 export const FirstMeet: React.FC = () => {
   const navigate = useNavigate();
@@ -31,7 +31,7 @@ export const FirstMeet: React.FC = () => {
       <div className="mt-6 flex-1 px-4">
         <DatePickerDrawer onConfirm={handleDateConfirm}>
           <Input
-            value={selectedDate ? formatDate(selectedDate) : ""}
+            value={selectedDate ? formatDateKorean(selectedDate) : ""}
             placeholder="날짜를 선택해주세요."
             status={selectedDate ? "active" : "inactive"}
             onClear={() => {
