--- src/shared/utils/date/FomatDateToKr.ts
@@ -1,10 +1,0 @@
-export const formatDateToKorean = (date: Date): string => {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-
-  const weekdays = ["일요일", "월요일", "화요일", "수요일", "목요일", "금요일", "토요일"];
-  const weekday = weekdays[date.getDay()];
-
-  return `${year}년 ${month}월 ${day}일 ${weekday}`;
-};
