+++ src/shared/utils/date/formatdate.ts
@@ -1,7 +1,35 @@
-// Drawer 안 날짜 형식 변환
-export const formatDate = (date: Date): string => {
+export const formatDateDot = (date: Date): string => {
+  // 2025. 06. 05
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  return `${year}. ${month}. ${day}`;
+};
+
+export const formatDateFull = (date: Date) => {
+  const yyyy = date.getFullYear();
+  const mm = String(date.getMonth() + 1).padStart(2, "0");
+  const dd = String(date.getDate()).padStart(2, "0");
+  const hh = String(date.getHours()).padStart(2, "0");
+  const min = String(date.getMinutes()).padStart(2, "0");
+
+  return `${yyyy}.${mm}.${dd} ${hh}:${min}`;
+};
+
+export const formatDateKorean = (date: Date): string => {
+  // 2025년 06월 05일
   const year = date.getFullYear();
   const month = String(date.getMonth() + 1).padStart(2, "0");
   const day = String(date.getDate()).padStart(2, "0");
   return `${year}년 ${month}월 ${day}일`;
+};
+
+export const formatDateKoreanWithWeekday = (date: Date): string => {
+  // 2025년 06월 05일 금요일
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  const weekdays = ["일요일", "월요일", "화요일", "수요일", "목요일", "금요일", "토요일"];
+  const weekday = weekdays[date.getDay()];
+  return `${year}년 ${month}월 ${day}일 ${weekday}`;
 };
